build(deps-dev): set `resolutions` to force latest `memfs` transitive dependency
The `page-with` package introduces a transitive dependency on `memfs`,
and specifies a semver range that excludes the latest version of `memfs`
which is compatible with Node v18.

Use the `resolutions` field of `package.json` to force Yarn classic
to install the (currently) latest version of `memfs` regardless of
semver range requested by `page-with`

This fixes at least one bug in `memfs` running in Node v18, which
caused excessive `console.error` spam which rightfully complained
of leaking file descriptors.

For docs on `resolutions` field behavior of Yarn classic, see:
https://classic.yarnpkg.com/en/docs/selective-version-resolutions