fix: infer attributes in non-mocked requests
Properties `withCredentials` and `responseType` are now transferred from
the patched instance to the actual request in non-mocked requests, so
the actual request takes into account these properties.

This also removes the piece of code which forwards responseType from the
actual request to the patched instance since the user agent will never
change the response type as per the XMLHttpRequest standard.