fix(xhr): do not pass response `Content-Encoding` to `TextDecoder` (#318)
- Fixes #308 

Thank you so much for working on this project, it is immensely useful!

TLDR: I am using the browser interceptors on a site that is serving
responses with `Content-Encoding: br` or `gzip` or `deflate` etc. This
breaks every request since `Content-Encoding` is typically for
compression algorithms, like `gzip`, like this site responds with, but
`TextDecoder` takes a character encoding, like `utf-8`. This breaks
every request even if nothing is done by the user because the library
logs the response, which implicitly calls `responseText()`. Avoiding
passing the `Content-Encoding` to `TextDecoder` fixes the issue.

---

Currently, the interceptors break on every request without me doing
anything because the `loadend` event logs the `response` which
implicitly calls `responseText()`:


https://github.com/mswjs/interceptors/blob/d8f785d2571a6941fd67e7d2f74c43db6aa528b5/src/interceptors/XMLHttpRequest/XMLHttpRequestOverride.ts#L447

The `responseText()` method passes the response' `Content-Encoding`
value to `decodeBuffer()` which passes it to `TextDecoder`.

However, it seems that the
[`Content-Encoding`](https://github.com/mswjs/interceptors/blob/main/src/interceptors/XMLHttpRequest/XMLHttpRequestOverride.ts#L483-L491)
header is conventionally for denoting the response' compression such as
`gzip`, `br`, `deflate` etc.

> Content encoding is mainly used to compress the message data without
losing information about the origin media type

See the [example
syntax](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Encoding#syntax)
and
[directives](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Encoding#directives).

Meanwhile, `TextDecoder` appears to be more concerned with [character
encoding](https://developer.mozilla.org/en-US/docs/Web/API/Encoding_API/Encodings),
such as `utf-8`.

For this reason, I recommend not passing through the `Content-Encoding`
value to `TextDecoder`, since it leads to errors on every request of
this form:

```
Uncaught RangeError: Failed to construct 'TextDecoder': The encoding label provided ('br') is invalid.
```

I confirmed that this change fixes the issue.

Once again thank you for all your great work on this project!

Co-authored-by: Artem Zakharchenko <kettanaito@gmail.com>