feat(XMLHttpRequest): use proxies to implement request interception (#337)
- Fixes #103
- Fixes #311
- Closes #310 (verified in 4724ae8dd4f606b16318b3b813153fff7cfb44d3)
- Fixes #309

It should be safe migrating to Proxies due to their [wide browser
support](https://caniuse.com/proxy).

## Changes

- **Uses ES6 Proxies to spy on `globalThis.XMLHttpRequest`.**
- Proxies `request.headers` changes of the Fetch API representation of
an XMLHttpRequest to update the underlying XHR instance correctly (in
the case the "request" listener has modified the request headers).
- Fixes the issue where a `readystatechange` event has been fired
_twice_ for XMLHttpRequest: once from the XHR instance itself, and the
second time from the underlying `http` implementation from JSDOM.
- Sets `credentials` to `same-origin` (previously `omit`) when the
`withCredentials` property of XMLHttpRequest is set to `false` or not
set.
- Deleting a request header in the "request" event listener will now
raise a warning for XMLHttpRequest. XHR headers cannot be deleted per
specification.
- **Spec-compliance:** We now throw `InvalidStateError` when reading the
request body improperly (e.g. using `responseText` when the
`responseType` is not an empty string or "text"). This also helps to
improve some of our tests as they were reading the response body of
XMLHttpRequest incorrectly.


## Roadmap

- [x] Add internal logger to `XMLHttpRequestController`
- [x] Emit the "response" event upon bypassed responses. There's also no
test for this, apparently.
- [x] Remove the previous implementation (`XMLHttpRequestOverride`).
- [x] Consider if setting the `X-Request-Id` on the bypassed requests is
still needed:

https://github.com/mswjs/interceptors/blob/d73624b222fd80a291486e13eb368d440bf21816/src/interceptors/XMLHttpRequest/XMLHttpRequestOverride.ts#L472-L477