fix(ClientRequest): stream response chunks as they come in (#373)
- Reported in
https://github.com/mswjs/msw/discussions/1464#discussioncomment-5789415

## Changes

Sets the `res` on `ClientRequest` instance _immediately_ (previously,
only when the response stream has finished reading).

Instead of pushing to the `IncomingMessage`, emit its `data` events for
each read chunk of the response stream. This is necessary because with a
simple `this.response.push(value)`, the end readable stream only
contained _the first chunk_. For some reason, any other chunks were
ignored with the previous implementation (they never emitted the `data`
event on `IncomingMessage`).

> This is potentially dangerous but let's see what the tests tell us. 

Also, we now mark the request as finished right before starting to
respond to it using a mocked response (previously, the request was
finished _after_ we were completely done with the response). This
respects the Node API better and also allows the consumer to modify the
intercepted request instance in the "request" event of the interceptor
_before_ responding to it.

Also, we no longer set `this.finished = true` on the Client Request. The
`finished` property has been deprecated in Node for some time, and we
should move away from it. We keep setting the `writableEnd = true` as
before, and we also set `writableFinished = true` since this change.